Get txes from local working with signing tool
