+++ pallets/template/src/common.rs
@@ -1,2 +1,2 @@
 /// The image id is unique per our program as it is a hash of some representation of the zkvm state. We can also obtain this from running the host code
-pub const TRANSFER_IMAGE_ID: [u32; 8] = [2486594279, 3606083639, 1255207922, 1378868077, 124846865, 115852313, 638666934, 300636237];
+pub const TRANSFER_IMAGE_ID: [u32; 8] = [2759512311, 937178575, 167584044, 2155138400, 2156841997, 4106496352, 776335246, 1022874217];
