+++ provers/transfer/transfer/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rollup-host"
+name = "prover-host"
 version = "0.1.0"
 edition = "2021"
 
@@ -12,4 +12,5 @@ subxt = { version = "0.25.0" }
 codec = { package = "parity-scale-codec", version = "3.2.2", default-features = false, features = ["derive"] }
 sp-core = { version = "7.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.39" }
 array-bytes = "6.1.0"
+clap = { version = "4.2.7", features = ["derive"] }
 serde_json = "1.0.96"
